requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
requests==2.32.0
python-nmap==0.7.1
reportlab==4.4.10
pillow>=10.0.0
mcp>=1.0.0
